requirements.txt
tornado==3.2
websocket-client
gevent
requests
motor
boom